models/Sign-VQ-Transformer-main/requirements.txt
lightning==2.1.4
numpy==1.24.2
tqdm==4.60.0
wandb==0.14.0
matplotlib==3.7.5
scikit-learn==1.2.2
seaborn==0.12.2
joblib==1.3.1
fastdtw==0.3.4
sacrebleu==2.4.0
scipy==1.10.1
tokenizers==0.15.1
jiwer==3.0.3
fasttext==0.9.2
opencv-python==4.9.0.80
mediapipe==0.10.9
pycpd==2.0.0
pickle5==0.0.11
pandas==2.0.3
ffmpeg==1.4
lmdb==1.4.1
